packages/mcp_servers/safepay_wallet/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "safepay-wallet-mcp"
version = "0.1.0"
description = "SafePay Wallet MCP Server for payment methods functionality"
authors = [
    { name = "SafePay Wallet Team" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "httpx>=0.26.0",
    "mcp>=0.1.0"
]
requires-python = ">=3.9"

[project.scripts]
safepay-wallet-mcp = "safepay_wallet.server:main"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto" 